+++ src/templates/vk-mini-app/src/pages/Splash/index.ts
@@ -1,3 +1,1 @@
-import Splash from './Splash';
-
-export default Splash;
+export { default } from './Splash';